Add 'mc' alias
'mc' copies a series of commits from a repo in a parent directory to the one in the current directory

This has a very specific use-case where a project contains a sub-project, where the sub-project maintains its own repo (that has not been added as a git submodule). Changes to the files in the sub-project need to be committed twice — once to the containing repo, and once to its own sub-project repo. The 'mc' alias allows us to quickly bring the commits into the sub-project repo.